refactor: Move hot compaction candidates to the hot module
