+++ compactor/src/compact.rs
@@ -83,7 +83,7 @@ pub type Result<T, E = Error> = std::result::Result<T, E>;
 #[derive(Debug)]
 pub struct Compactor {
     /// Shards assigned to this compactor
-    shards: Vec<ShardId>,
+    pub(crate) shards: Vec<ShardId>,
 
     /// Object store for reading and persistence of parquet files
     pub(crate) store: ParquetStorage,
@@ -104,7 +104,7 @@ pub struct Compactor {
     pub(crate) config: CompactorConfig,
 
     /// Gauge for the number of compaction partition candidates before filtering
-    compaction_candidate_gauge: Metric<U64Gauge>,
+    pub(crate) compaction_candidate_gauge: Metric<U64Gauge>,
 
     /// Gauge for the number of Parquet file candidates after filtering. The recorded values have
     /// attributes for the compaction level of the file and whether the file was selected for
@@ -250,113 +250,6 @@ impl Compactor {
         }
     }
 
-    /// Return a list of the most recent highest ingested throughput partitions.
-    /// The highest throughput partitions are prioritized as follows:
-    ///  1. If there are partitions with new ingested files within the last 4 hours, pick them.
-    ///  2. If no new ingested files in the last 4 hours, will look for partitions with new writes
-    ///     within the last 24 hours.
-    ///  3. If there are no ingested files within the last 24 hours, will look for partitions
-    ///     with any new ingested files in the past.
-    ///
-    /// * New ingested files means non-deleted L0 files
-    /// * In all cases above, for each shard, N partitions with the most new ingested files
-    ///   will be selected and the return list will include at most, P = N * S, partitions where S
-    ///   is the number of shards this compactor handles.
-    pub async fn hot_partitions_to_compact(
-        &self,
-        // Max number of the most recent highest ingested throughput partitions
-        // per shard we want to read
-        max_num_partitions_per_shard: usize,
-        // Minimum number of the most recent writes per partition we want to count
-        // to prioritize partitions
-        min_recent_ingested_files: usize,
-    ) -> Result<Vec<Arc<PartitionCompactionCandidateWithInfo>>> {
-        let compaction_type = "hot";
-        let mut candidates = Vec::with_capacity(self.shards.len() * max_num_partitions_per_shard);
-
-        for shard_id in &self.shards {
-            let attributes = Attributes::from([
-                ("shard_id", format!("{}", *shard_id).into()),
-                ("partition_type", compaction_type.into()),
-            ]);
-            let mut repos = self.catalog.repositories().await;
-
-            // Get the most recent highest ingested throughput partitions within
-            // the last 4 hours. If not, increase to 24 hours
-            let mut num_partitions = 0;
-            for num_hours in [4, 24] {
-                let time_at_num_hours_ago =
-                    Timestamp::from(self.time_provider.hours_ago(num_hours));
-
-                let mut partitions = repos
-                    .parquet_files()
-                    .recent_highest_throughput_partitions(
-                        *shard_id,
-                        time_at_num_hours_ago,
-                        min_recent_ingested_files,
-                        max_num_partitions_per_shard,
-                    )
-                    .await
-                    .context(HighestThroughputPartitionsSnafu {
-                        shard_id: *shard_id,
-                    })?;
-
-                if !partitions.is_empty() {
-                    debug!(
-                        shard_id = shard_id.get(),
-                        num_hours,
-                        n = partitions.len(),
-                        "found high-throughput partitions"
-                    );
-                    num_partitions = partitions.len();
-                    candidates.append(&mut partitions);
-                    break;
-                }
-            }
-
-            // Record metric for candidates per shard
-            debug!(
-                shard_id = shard_id.get(),
-                n = num_partitions,
-                compaction_type,
-                "compaction candidates",
-            );
-            let number_gauge = self.compaction_candidate_gauge.recorder(attributes);
-            number_gauge.set(num_partitions as u64);
-        }
-
-        // Get extra needed information for selected partitions
-        let start_time = self.time_provider.now();
-
-        // Column types and their counts of the tables of the partition candidates
-        debug!(
-            num_candidates=?candidates.len(),
-            compaction_type,
-            "start getting column types for the partition candidates"
-        );
-        let table_columns = self.table_columns(&candidates).await?;
-
-        // Add other compaction-needed info into selected partitions
-        debug!(
-            num_candidates=?candidates.len(),
-            compaction_type,
-            "start getting additional info for the partition candidates"
-        );
-        let candidates = self
-            .add_info_to_partitions(&candidates, &table_columns)
-            .await?;
-
-        if let Some(delta) = self.time_provider.now().checked_duration_since(start_time) {
-            let attributes = Attributes::from(&[("partition_type", compaction_type)]);
-            let duration = self
-                .partitions_extra_info_reading_duration
-                .recorder(attributes);
-            duration.record(delta);
-        }
-
-        Ok(candidates)
-    }
-
     /// Return a list of partitions that:
     ///
     /// - Have not received any writes in 8 hours (determined by all level 0 and level 1 parquet
@@ -439,7 +332,7 @@ impl Compactor {
     }
 
     /// Get column types for tables of given partitions
-    async fn table_columns(
+    pub(crate) async fn table_columns(
         &self,
         partitions: &[PartitionParam],
     ) -> Result<HashMap<TableId, Vec<ColumnTypeCount>>> {
@@ -460,7 +353,7 @@ impl Compactor {
     }
 
     /// Add namespace and table information to partition candidates.
-    async fn add_info_to_partitions(
+    pub(crate) async fn add_info_to_partitions(
         &self,
         partitions: &[PartitionParam],
         table_columns: &HashMap<TableId, Vec<ColumnTypeCount>>,
@@ -706,254 +599,6 @@ pub mod tests {
         ];
         let bytes = estimate_arrow_bytes_for_file(&columns, row_count);
         assert_eq!(bytes, 12979); // 880 + 1088 + 11000 + 11
-    }
-
-    #[tokio::test]
-    async fn test_hot_partitions_to_compact() {
-        let catalog = TestCatalog::new();
-
-        // Create a db with 2 shards, one with 4 empty partitions and the other one with one
-        // empty partition
-        let mut txn = catalog.catalog.start_transaction().await.unwrap();
-
-        let topic = txn.topics().create_or_get("foo").await.unwrap();
-        let pool = txn.query_pools().create_or_get("foo").await.unwrap();
-        let namespace = txn
-            .namespaces()
-            .create(
-                "namespace_hot_partitions_to_compact",
-                "inf",
-                topic.id,
-                pool.id,
-            )
-            .await
-            .unwrap();
-        let table = txn
-            .tables()
-            .create_or_get("test_table", namespace.id)
-            .await
-            .unwrap();
-        let shard = txn
-            .shards()
-            .create_or_get(&topic, ShardIndex::new(1))
-            .await
-            .unwrap();
-        let partition1 = txn
-            .partitions()
-            .create_or_get("one".into(), shard.id, table.id)
-            .await
-            .unwrap();
-        let partition2 = txn
-            .partitions()
-            .create_or_get("two".into(), shard.id, table.id)
-            .await
-            .unwrap();
-        let partition3 = txn
-            .partitions()
-            .create_or_get("three".into(), shard.id, table.id)
-            .await
-            .unwrap();
-        let partition4 = txn
-            .partitions()
-            .create_or_get("four".into(), shard.id, table.id)
-            .await
-            .unwrap();
-        // other shard
-        let another_table = txn
-            .tables()
-            .create_or_get("another_test_table", namespace.id)
-            .await
-            .unwrap();
-        let another_shard = txn
-            .shards()
-            .create_or_get(&topic, ShardIndex::new(2))
-            .await
-            .unwrap();
-        let another_partition = txn
-            .partitions()
-            .create_or_get(
-                "another_partition".into(),
-                another_shard.id,
-                another_table.id,
-            )
-            .await
-            .unwrap();
-        // update sort key for this another_partition
-        let another_partition = txn
-            .partitions()
-            .update_sort_key(another_partition.id, &["tag1", "time"])
-            .await
-            .unwrap();
-        txn.commit().await.unwrap();
-
-        // Create a compactor
-        let time_provider = Arc::new(SystemProvider::new());
-        let config = make_compactor_config();
-        let compactor = Compactor::new(
-            vec![shard.id, another_shard.id],
-            Arc::clone(&catalog.catalog),
-            ParquetStorage::new(Arc::clone(&catalog.object_store)),
-            Arc::new(Executor::new(1)),
-            time_provider,
-            BackoffConfig::default(),
-            config,
-            Arc::new(metric::Registry::new()),
-        );
-
-        // Some times in the past to set to created_at of the files
-        let time_now = Timestamp::from(compactor.time_provider.now());
-        let time_three_minutes_ago = Timestamp::from(compactor.time_provider.minutes_ago(3));
-        let time_five_hour_ago = Timestamp::from(compactor.time_provider.hours_ago(5));
-        let time_38_hour_ago = Timestamp::from(compactor.time_provider.hours_ago(38));
-
-        // Basic parquet info
-        let p1 = ParquetFileParams {
-            shard_id: shard.id,
-            namespace_id: namespace.id,
-            table_id: table.id,
-            partition_id: partition1.id,
-            object_store_id: Uuid::new_v4(),
-            max_sequence_number: SequenceNumber::new(100),
-            min_time: Timestamp::new(1),
-            max_time: Timestamp::new(5),
-            file_size_bytes: 1337,
-            row_count: 0,
-            compaction_level: CompactionLevel::Initial, // level of file of new writes
-            created_at: time_now,
-            column_set: ColumnSet::new([ColumnId::new(1), ColumnId::new(2)]),
-        };
-
-        // Note: The order of the test cases below is important and should not be changed
-        // because they depend on the order of the writes and their content. For example,
-        // in order to test `Case 3`, we do not need to add asserts for `Case 1` and `Case 2`,
-        // but all the writes, deletes and updates in Cases 1 and 2 are a must for testing Case 3.
-        // In order words, the last Case needs all content of previous tests.
-        // This shows the priority of selecting compaction candidates
-
-        // --------------------------------------
-        // Case 1: no files yet --> no partition candidates
-        //
-        let candidates = compactor.hot_partitions_to_compact(1, 1).await.unwrap();
-        assert!(candidates.is_empty());
-
-        // --------------------------------------
-        // Case 2: no non-deleleted L0 files -->  no partition candidates
-        //
-        // partition1 has a deleted L0
-        let mut txn = catalog.catalog.start_transaction().await.unwrap();
-        let pf1 = txn.parquet_files().create(p1.clone()).await.unwrap();
-        txn.parquet_files().flag_for_delete(pf1.id).await.unwrap();
-        //
-        // partition2 has a non-L0 file
-        let p2 = ParquetFileParams {
-            object_store_id: Uuid::new_v4(),
-            partition_id: partition2.id,
-            ..p1.clone()
-        };
-        let pf2 = txn.parquet_files().create(p2).await.unwrap();
-        txn.parquet_files()
-            .update_compaction_level(&[pf2.id], CompactionLevel::FileNonOverlapped)
-            .await
-            .unwrap();
-        txn.commit().await.unwrap();
-        // No non-deleted level 0 files yet --> no candidates
-        let candidates = compactor.hot_partitions_to_compact(1, 1).await.unwrap();
-        assert!(candidates.is_empty());
-
-        // --------------------------------------
-        // Case 3: no new recent writes (within the last 24 hours) --> no partition candidates
-        // (the cold case will pick them up)
-        //
-        // partition2 has an old (more than 8 hours ago) non-deleted level 0 file
-        let mut txn = catalog.catalog.start_transaction().await.unwrap();
-        let p3 = ParquetFileParams {
-            object_store_id: Uuid::new_v4(),
-            partition_id: partition2.id,
-            created_at: time_38_hour_ago,
-            ..p1.clone()
-        };
-        let _pf3 = txn.parquet_files().create(p3).await.unwrap();
-        txn.commit().await.unwrap();
-
-        // No hot candidates
-        let candidates = compactor.hot_partitions_to_compact(1, 1).await.unwrap();
-        assert!(candidates.is_empty());
-
-        // --------------------------------------
-        // Case 4: has one partition with recent writes (5 hours ago) --> return that partition
-        //
-        // partition4 has a new write 5 hours ago
-        let mut txn = catalog.catalog.start_transaction().await.unwrap();
-        let p4 = ParquetFileParams {
-            object_store_id: Uuid::new_v4(),
-            partition_id: partition4.id,
-            created_at: time_five_hour_ago,
-            ..p1.clone()
-        };
-        let _pf4 = txn.parquet_files().create(p4).await.unwrap();
-        txn.commit().await.unwrap();
-        //
-        // Has at least one partition with a recent write --> make it a candidate
-        let candidates = compactor.hot_partitions_to_compact(1, 1).await.unwrap();
-        assert_eq!(candidates.len(), 1);
-        assert_eq!(candidates[0].id(), partition4.id);
-
-        // --------------------------------------
-        // Case 5: has 2 partitions with 2 different groups of recent writes:
-        //  1. Within the last 4 hours
-        //  2. Within the last 24 hours but older than 4 hours ago
-        // When we have group 1, we will ignore partitions in group 2
-        //
-        // partition3 has a new write 3 hours ago
-        let mut txn = catalog.catalog.start_transaction().await.unwrap();
-        let p5 = ParquetFileParams {
-            object_store_id: Uuid::new_v4(),
-            partition_id: partition3.id,
-            created_at: time_three_minutes_ago,
-            ..p1.clone()
-        };
-        let _pf5 = txn.parquet_files().create(p5).await.unwrap();
-        txn.commit().await.unwrap();
-        //
-        // make partitions in the most recent group candidates
-        let candidates = compactor.hot_partitions_to_compact(1, 1).await.unwrap();
-        assert_eq!(candidates.len(), 1);
-        assert_eq!(candidates[0].id(), partition3.id);
-
-        // --------------------------------------
-        // Case 6: has partition candidates for 2 shards
-        //
-        // The another_shard now has non-deleted level-0 file ingested 5 hours ago
-        let mut txn = catalog.catalog.start_transaction().await.unwrap();
-        let p6 = ParquetFileParams {
-            object_store_id: Uuid::new_v4(),
-            shard_id: another_shard.id,
-            table_id: another_table.id,
-            partition_id: another_partition.id,
-            created_at: time_five_hour_ago,
-            ..p1.clone()
-        };
-        let _pf6 = txn.parquet_files().create(p6).await.unwrap();
-        txn.commit().await.unwrap();
-        //
-        // Will have 2 candidates, one for each shard
-        let mut candidates = compactor.hot_partitions_to_compact(1, 1).await.unwrap();
-        candidates.sort_by_key(|c| c.candidate);
-        assert_eq!(candidates.len(), 2);
-
-        assert_eq!(candidates[0].id(), partition3.id);
-        assert_eq!(candidates[0].shard_id(), shard.id);
-        assert_eq!(*candidates[0].namespace, namespace);
-        assert_eq!(*candidates[0].table, table);
-        assert_eq!(candidates[0].partition_key, partition3.partition_key);
-        assert_eq!(candidates[0].sort_key, partition3.sort_key()); // this sort key is None
-
-        assert_eq!(candidates[1].id(), another_partition.id);
-        assert_eq!(candidates[1].shard_id(), another_shard.id);
-        assert_eq!(*candidates[1].namespace, namespace);
-        assert_eq!(*candidates[1].table, another_table);
-        assert_eq!(candidates[1].partition_key, another_partition.partition_key);
-        assert_eq!(candidates[1].sort_key, another_partition.sort_key()); // this sort key is Some(tag1, time)
     }
 
     fn make_compactor_config() -> CompactorConfig {
